hyperliquid-agent/pyproject.toml
[project]
name = "nxfh01-agent"
version = "0.1.0"
description = "NXFH01: Autonomous Hyperliquid perp trading agent for the Virtuals Degen Claw competition"
requires-python = ">=3.11"

dependencies = [
    "hyperliquid-python-sdk>=0.22.0",
    "game-sdk>=0.1.0",
    "eth-account>=0.13.0",
    "pydantic>=2.5",
    "pydantic-settings>=2.0",
    "python-dotenv>=1.0",
    "pyyaml>=6.0",
    "pandas>=2.0",
    "ta>=0.11",
    "httpx>=0.27",
]

[project.optional-dependencies]
dev = ["pytest>=8.0", "pytest-asyncio>=0.23", "ruff>=0.4"]

[project.scripts]
nxfh01 = "src.main:main"

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]

[tool.ruff]
line-length = 100
target-version = "py311"
